+++ src/browser.ts
@@ -1,9 +1,10 @@
 import { setWebConfig, webAnalytics, getCountryCode, webData, recordEvent } from "./code/web";
 import { RecordEventInteractionInput, RecordEventPageViewInput, StatsDevice, StatsEventType } from "./types/stats";
-import { WebConfig } from "./types/web";
+import { RecordCallback, WebConfig } from "./types/web";
 
 export {
     WebConfig,
+    RecordCallback,
     RecordEventPageViewInput,
     RecordEventInteractionInput,
     StatsDevice,